Revert "Use int64 for collecting command stats"

Closes omnibus-gitlab#7321

See merge request https://gitlab.com/gitlab-org/gitaly/-/merge_requests/5074

Merged-by: Patrick Steinhardt <psteinhardt@gitlab.com>
Approved-by: Quang-Minh Nguyen <qmnguyen@gitlab.com>
Approved-by: Patrick Steinhardt <psteinhardt@gitlab.com>

+++ internal/middleware/commandstatshandler/commandstatshandler_test.go
@@ -113,7 +113,7 @@ func TestInterceptor(t *testing.T) {
 				//
 				// This test will break again as soon as we bump the minimum Git
 				// version and thus remove the version check.
-				"command.count": int64(2),
+				"command.count": 2,
 			},
 		},
 		{
@@ -133,7 +133,7 @@ func TestInterceptor(t *testing.T) {
 				}
 			},
 			expectedLogData: map[string]interface{}{
-				"command.count": int64(1),
+				"command.count": 1,
 			},
 		},
 	}
@@ -166,5 +166,5 @@ func TestFieldsProducer(t *testing.T) {
 	stats := command.StatsFromContext(ctx)
 	stats.RecordMax("stub", 42)
 
-	require.Equal(t, logrus.Fields{"stub": int64(42)}, FieldsProducer(ctx))
+	require.Equal(t, logrus.Fields{"stub": 42}, FieldsProducer(ctx))
 }
